Move GeyserRelocationMixinExtension to fabric only
