+++ fabric/src/main/java/org/geysermc/hydraulic/fabric/HydraulicFabricMod.java
@@ -1,5 +1,6 @@
 package org.geysermc.hydraulic.fabric;
 
+import com.llamalad7.mixinextras.utils.MixinInternals;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import org.geysermc.hydraulic.HydraulicImpl;
@@ -11,6 +12,9 @@ public class HydraulicFabricMod implements ModInitializer {
 
     @Override
     public void onInitialize() {
+        // Register the GeyserRelocationMixinExtension
+        MixinInternals.registerExtension(new GeyserRelocationMixinExtension());
+
         this.hydraulic = HydraulicImpl.load(HydraulicPlatform.FABRIC, new HydraulicFabricBootstrap());
 
         ServerLifecycleEvents.SERVER_STARTING.register(this.hydraulic::onServerStarting);
